Refactor cmc pro service to use services pattern.
