+++ lib/scheduled_jobs/update_cmc_prices.rb
@@ -15,15 +15,15 @@ class UpdateCmcPrices
   def perform_prices(id)
     case id
     when 1 # Top 1-100
-      CoinMarketCapProService.new.ticker_update(1, 100)
+      CoinMarketCapPro::UpdateTickerService.call(1, 100)
     when 2 # Top 101-200
-      CoinMarketCapProService.new.ticker_update(101, 100)
+      CoinMarketCapPro::UpdateTickerService.call(101, 100)
     when 3 # Top 201-500
-      CoinMarketCapProService.new.ticker_update(201, 300)
+      CoinMarketCapPro::UpdateTickerService.call(201, 300)
     when 4 # Top 501-1000
-      CoinMarketCapProService.new.ticker_update(501, 500)
+      CoinMarketCapPro::UpdateTickerService.call(501, 500)
     when 5 # Top 1001+
-      CoinMarketCapProService.new.ticker_update(1001, 5000) # 5000 is CMC's max limit
+      CoinMarketCapPro::UpdateTickerService.call(1001, 5000) # 5000 is CMC's max limit
     end
   end
 end
